drive_bot_web/requirements.txt
APScheduler==3.6.3
billiard==3.6.4.0
celery==5.2.3
Django==4.2
django-admin-list-filter-dropdown==1.0.2
django-celery-results==2.2.0
django-celerybeat-status==0.0.10
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-environ==0.4.5
django-json-widget==1.1.1
django-mptt==0.14.0
ipython==8.8.0
kombu==5.2.3
uwsgi==2.0.20
psycopg2-binary==2.9.3
redis==3.3.11
python-telegram-bot==20.6
telegram_django_bot==1.1.0
djangorestframework==3.14
django-filter==23.3
